Restructure graph `util.ts`
