+++ libs/@blockprotocol/graph/src/shared/util/must-be-defined.ts
@@ -1,0 +1,7 @@
+export const mustBeDefined = <T>(x: T | undefined, message?: string): T => {
+  if (x === undefined) {
+    throw new Error(`invariant was broken: ${message ?? ""}`);
+  }
+
+  return x;
+};
