+++ libs/@blockprotocol/graph/src/shared/util/subtype.ts
@@ -1,0 +1,8 @@
+/**
+ * A somewhat satisfactory SubType helper type. This provides _some_ level of type safety when trying to mark one
+ * type as constraining another type.
+ */
+export type Subtype<
+  Type extends Record<string, unknown>,
+  Sub extends Type,
+> = Sub;
